# internal/server/api/spec/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "wdw-fleet",
    "description": "API-first vehicle fleet management service (FleetAware).",
    "version": "0.1.0-draft",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
...
  "paths": {
...
    "/openapi.json": {
      "get": {
        "operationId": "getOpenAPIJSON",
        "summary": "The OpenAPI 3.1 contract, JSON serialization.",
        "tags": [
          "System"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "default": {
            "description": "Error (RFC 7807 problem+json).",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "x-wdw-auth": "public"
      }
    },
    "/openapi.yaml": {
      "get": {
        "operationId": "getOpenAPIYAML",
        "summary": "The OpenAPI 3.1 contract, YAML serialization.",
        "tags": [
          "System"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "default": {
            "description": "Error (RFC 7807 problem+json).",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "x-wdw-auth": "public"
      }
    },
    "/v1/vehicles": {
      "get": {
        "operationId": "listVehicles",
        "summary": "List every non-deleted vehicle in the fleet.",
        "tags": [
          "Fleet"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "items": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "acquisition_cost_cents": {
                            "type": "integer",
                            "format": "int32",
                            "nullable": true
                          },
                          "acquisition_date": {
                            "type": "string",
                            "format": "date",
                            "nullable": true
                          },
                          "created_at": {
                            "type": "string",
                            "format": "date-time",
                            "readOnly": true
                          },
                          "custom_fields": {
                            "type": "object",
                            "additionalProperties": true
                          },
                          "glyph": {
                            "type": "string",
                            "enum": [
                              "truck",
                              "tractor",
                              "atv",
                              "trailer",
                              "gen",
                              "car",
                              "track",
                              "race"
                            ]
                          },
                          "id": {
                            "type": "string",
                            "format": "uuid",
                            "readOnly": true
                          },
                          "license_plate": {
                            "type": "string"
                          },
                          "make": {
                            "type": "string"
                          },
                          "model": {
                            "type": "string"
                          },
                          "name": {
                            "type": "string"
                          },
                          "notes": {
                            "type": "string"
                          },
                          "odometer": {
                            "type": "integer",
                            "format": "int32"
                          },
                          "odometer_unit": {
                            "type": "string",
                            "enum": [
                              "mi",
                              "km",
                              "hr",
                              "events"
                            ]
                          },
                          "photo_url": {
                            "type": "string"
                          },
                          "sale_date": {
                            "type": "string",
                            "format": "date",
                            "nullable": true
                          },
                          "sale_price_cents": {
                            "type": "integer",
                            "format": "int32",
                            "nullable": true
                          },
                          "status": {
                            "type": "string",
                            "enum": [
                              "active",
                              "inactive",
                              "sold"
                            ]
                          },
                          "tags": {
                            "type": "array",
                            "items": {
                              "type": "string"
                            }
                          },
                          "updated_at": {
                            "type": "string",
                            "format": "date-time",
                            "readOnly": true
                          },
                          "vin": {
                            "type": "string"
                          },
                          "year": {
                            "type": "integer",
                            "format": "int32",
                            "nullable": true
                          }
                        },
                        "required": [
                          "id",
                          "name",
                          "make",
                          "model",
                          "vin",
                          "license_plate",
                          "glyph",
                          "odometer",
                          "odometer_unit",
                          "status",
                          "notes",
                          "custom_fields",
                          "photo_url",
                          "tags",
                          "created_at",
                          "updated_at"
                        ]
                      }
                    }
                  },
                  "required": [
                    "items"
                  ]
                }
              }
            }
          },
          "default": {
            "description": "Error (RFC 7807 problem+json).",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "x-wdw-auth": "public"
      },
      "post": {
        "operationId": "createVehicle",
        "summary": "Create a vehicle.",
        "tags": [
          "Fleet"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "acquisition_cost_cents": {
                    "type": "integer",
                    "format": "int32",
                    "nullable": true
                  },
                  "acquisition_date": {
                    "type": "string",
                    "format": "date",
                    "nullable": true
                  },
                  "custom_fields": {
                    "type": "object",
                    "additionalProperties": true
                  },
                  "glyph": {
                    "type": "string",
                    "enum": [
                      "truck",
                      "tractor",
                      "atv",
                      "trailer",
                      "gen",
                      "car",
                      "track",
                      "race"
                    ]
                  },
                  "license_plate": {
                    "type": "string"
                  },
                  "make": {
                    "type": "string"
                  },
                  "model": {
                    "type": "string"
                  },
                  "name": {
                    "type": "string"
                  },
                  "notes": {
                    "type": "string"
                  },
                  "odometer": {
                    "type": "integer",
                    "format": "int32",
                    "nullable": true
                  },
                  "odometer_unit": {
                    "type": "string",
                    "enum": [
                      "mi",
                      "km",
                      "hr",
                      "events"
                    ]
                  },
                  "status": {
                    "type": "string",
                    "enum": [
                      "active",
                      "inactive",
                      "sold"
                    ]
                  },
                  "tags": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "vin": {
                    "type": "string"
                  },
                  "year": {
                    "type": "integer",
                    "format": "int32",
                    "nullable": true
                  }
                },
                "required": [
                  "name"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "acquisition_cost_cents": {
                      "type": "integer",
                      "format": "int32",
                      "nullable": true
                    },
                    "acquisition_date": {
                      "type": "string",
                      "format": "date",
                      "nullable": true
                    },
                    "created_at": {
                      "type": "string",
                      "format": "date-time",
                      "readOnly": true
                    },
                    "custom_fields": {
                      "type": "object",
                      "additionalProperties": true
                    },
                    "glyph": {
                      "type": "string",
                      "enum": [
                        "truck",
                        "tractor",
                        "atv",
                        "trailer",
                        "gen",
                        "car",
                        "track",
                        "race"
                      ]
                    },
                    "id": {
                      "type": "string",
                      "format": "uuid",
                      "readOnly": true
                    },
                    "license_plate": {
                      "type": "string"
                    },
                    "make": {
                      "type": "string"
                    },
                    "model": {
                      "type": "string"
                    },
                    "name": {
                      "type": "string"
                    },
                    "notes": {
                      "type": "string"
                    },
                    "odometer": {
                      "type": "integer",
                      "format": "int32"
                    },
                    "odometer_unit": {
                      "type": "string",
                      "enum": [
                        "mi",
                        "km",
                        "hr",
                        "events"
                      ]
                    },
                    "photo_url": {
                      "type": "string"
                    },
                    "sale_date": {
                      "type": "string",
                      "format": "date",
                      "nullable": true
                    },
                    "sale_price_cents": {
                      "type": "integer",
                      "format": "int32",
                      "nullable": true
                    },
                    "status": {
                      "type": "string",
                      "enum": [
                        "active",
                        "inactive",
                        "sold"
                      ]
                    },
                    "tags": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    },
                    "updated_at": {
                      "type": "string",
                      "format": "date-time",
                      "readOnly": true
                    },
                    "vin": {
                      "type": "string"
                    },
                    "year": {
                      "type": "integer",
                      "format": "int32",
                      "nullable": true
                    }
                  },
                  "required": [
                    "id",
                    "name",
                    "make",
                    "model",
                    "vin",
                    "license_plate",
                    "glyph",
                    "odometer",
                    "odometer_unit",
                    "status",
                    "notes",
                    "custom_fields",
                    "photo_url",
                    "tags",
                    "created_at",
                    "updated_at"
                  ]
                }
              }
            }
          },
          "default": {
            "description": "Error (RFC 7807 problem+json).",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "x-wdw-auth": "public"
      }
    },
...
  },
  "components": {
    "schemas": {
      "Problem": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "description": "Machine-readable error code."
          },
          "detail": {
            "type": "string",
            "description": "Human-readable explanation of this occurrence."
          },
          "instance": {
            "type": "string",
            "format": "uri",
            "description": "URI identifying this occurrence."
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "description": "HTTP status code."
          },
          "title": {
            "type": "string",
            "description": "Short human-readable summary."
          },
          "type": {
            "type": "string",
            "format": "uri",
            "description": "Problem type URI."
          }
        },
        "required": [
          "title",
          "status",
          "code"
        ]
      }
    }
  }
}